# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api-json":{"get":{"operationId":"AppController_code","summary":"获取接口文档json数据","parameters":[],"responses":{"200":{"description":""}}}},"/api/user/init":{"get":{"operationId":"UserController_initData","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/code":{"get":{"operationId":"UserController_code","summary":"获取图文验证码","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/emailCode":{"get":{"operationId":"UserController_emailCode","summary":"获取邮箱验证码","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/qrCode":{"get":{"operationId":"UserController_qrCode","summary":"获取微信登录二维码","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user":{"post":{"operationId":"UserController_create","summary":"创建用户","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]},"get":{"operationId":"UserController_findAll","summary":"查找所有用户","parameters":[{"name":"keyWord","required":false,"in":"query","description":"关键字","schema":{"type":"string"}},{"name":"pageSize","required":false,"in":"query","description":"每页条数","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","description":"页码","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/{id}":{"get":{"operationId":"UserController_findOne","summary":"根据id查找用户","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]},"patch":{"operationId":"UserController_update","summary":"修改用户","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"]},"delete":{"operationId":"UserController_remove","summary":"删除用户","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/roleMenu/{id}":{"get":{"operationId":"UserController_findMenuByRuleId","summary":"根据ruleId查找菜单","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/getRole/role":{"get":{"operationId":"UserController_GetRole","summary":"查询角色","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","summary":"用户登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}},"/api/auth/code":{"get":{"operationId":"AuthController_code","summary":"获取图文验证码","parameters":[],"responses":{"200":{"description":""}}}},"/api/menu":{"post":{"operationId":"MenuController_menu","summary":"新增/修改菜单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"201":{"description":""}},"tags":["菜单"]},"get":{"operationId":"MenuController_findAll","summary":"获取菜单信息","parameters":[],"responses":{"200":{"description":""}},"tags":["菜单"]}},"/api/menu/queryMenuRouters":{"get":{"operationId":"MenuController_queryMenuRouters","summary":"获取所有菜单","parameters":[],"responses":{"200":{"description":""}},"tags":["菜单"]}},"/api/upload":{"post":{"operationId":"UploadController_upload","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"UploadController_getAllFile","parameters":[],"responses":{"200":{"description":""}}}},"/api/upload/addFile":{"post":{"operationId":"UploadController_create","parameters":[],"responses":{"201":{"description":""}}}},"/api/upload/{id}":{"get":{"operationId":"UploadController_getFile","parameters":[],"responses":{"200":{"description":""}}},"delete":{"operationId":"UploadController_deleteFile","parameters":[],"responses":{"200":{"description":""}}}},"/api/upload/download/{id}":{"get":{"operationId":"UploadController_export","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"nest-shop","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"id":{"type":"number","description":"用户id","example":1},"userName":{"type":"string","description":"用户名","example":"admin"},"password":{"type":"string","description":"密码","example":"123456"},"email":{"type":"string","description":"邮箱","example":"1640551913@qq.com"},"role":{"type":"string","description":"角色","example":"editor"},"phone":{"type":"string","description":"手机号","example":"13888888888"},"address":{"type":"string","description":"地址","example":"广州"},"state":{"type":"number","description":"状态","example":1}},"required":["id","userName","password","email","role","phone","address","state"]},"UpdateUserDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{"userName":{"type":"string","description":"用户名","example":"admin"},"email":{"type":"string","description":"邮箱","example":"1640551913@qq.com"},"password":{"type":"string","description":"密码","example":"123456"},"code":{"type":"string","description":"验证码"}},"required":["userName","email","password","code"]},"CreateMenuDto":{"type":"object","properties":{"menuId":{"type":"number","description":"菜单id","example":1},"parentMenuId":{"type":"number","description":"父菜单id","example":1},"title":{"type":"string","description":"菜单名"},"path":{"type":"string","description":"URL路径"},"name":{"type":"string","description":"标题"},"component":{"type":"string","description":"组件路径"},"redirect":{"type":"string","description":"重定向"},"icon":{"type":"string","description":"图标"},"hidden":{"type":"boolean","description":"是否隐藏"}},"required":["menuId","parentMenuId","title","path","name","component","redirect","icon","hidden"]}}},"externalDocs":{"description":"api-json","url":"http://localhost:3000/api-json"}}